+++ docs/conf.py
@@ -15,7 +15,7 @@ sys.path.insert(0, os.path.abspath(".."))
 project = "maize-contrib"
 copyright = "2023, Molecular AI Group"
 author = "Thomas Löhr"
-release = "0.2.2"
+release = "0.2.3"
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration