+++ maize/graphs/mai/dock.py
@@ -1,5 +1,6 @@
 """Docking workflows and subgraphs"""
 
+from collections.abc import Callable
 import json
 from pathlib import Path
 import sys